+++ src/datalevin/scan.clj
@@ -42,6 +42,83 @@
         (when-not (l/writing? lmdb)
           (i/return-rtx lmdb rtx))))))
 
+(defn get-rank
+  [lmdb dbi-name k k-type]
+  (i/check-ready lmdb)
+  (let [dbi (i/get-dbi lmdb dbi-name false)
+        rtx (if (l/writing? lmdb)
+              @(l/write-txn lmdb)
+              (i/get-rtx lmdb))]
+    (try
+      (l/put-key rtx k k-type)
+      (l/get-key-rank dbi rtx)
+      (catch Throwable e
+        (raise "Fail to get-rank: " e
+               {:dbi dbi-name :k k :k-type k-type}))
+      (finally
+        (when-not (l/writing? lmdb)
+          (i/return-rtx lmdb rtx))))))
+
+(defn get-by-rank
+  [lmdb dbi-name rank k-type v-type ignore-key?]
+  (i/check-ready lmdb)
+  (let [dbi (i/get-dbi lmdb dbi-name false)
+        rtx (if (l/writing? lmdb)
+              @(l/write-txn lmdb)
+              (i/get-rtx lmdb))]
+    (try
+      (when-let [[kb vb] (l/get-key-by-rank dbi rtx rank)]
+        (if ignore-key?
+          (b/read-buffer vb v-type)
+          [(b/read-buffer kb k-type) (b/read-buffer vb v-type)]))
+      (catch Throwable e
+        (raise "Fail to get-by-rank: " e
+               {:dbi dbi-name :rank rank :k-type k-type :v-type v-type}))
+      (finally
+        (when-not (l/writing? lmdb)
+          (i/return-rtx lmdb rtx))))))
+
+(defn sample-kv
+  [lmdb dbi-name n k-type v-type ignore-key?]
+  (i/check-ready lmdb)
+  (let [total     (i/entries lmdb dbi-name)
+        indices   (u/reservoir-sampling total n)
+        list-dbi? (i/list-dbi? lmdb dbi-name)]
+    (when indices
+      (let [dbi (i/get-dbi lmdb dbi-name false)
+            rtx (if (l/writing? lmdb)
+                  @(l/write-txn lmdb)
+                  (i/get-rtx lmdb))
+            cur (l/get-cursor dbi rtx)]
+        (try
+          (with-open [^AutoCloseable iter
+                      (.iterator
+                        ^Iterable (if list-dbi?
+                                    (l/iterate-list-sample
+                                      dbi rtx cur indices 0 0 [:all] k-type)
+                                    (l/iterate-key-sample
+                                      dbi rtx cur indices 0 0 [:all] k-type)))]
+            (let [^SpillableVector holder
+                  (sp/new-spillable-vector nil (:spill-opts (i/env-opts lmdb)))]
+              (loop []
+                (if (.hasNext ^Iterator iter)
+                  (let [kv (.next ^Iterator iter)
+                        v  (b/read-buffer (l/v kv) v-type)]
+                    (.cons holder (if ignore-key?
+                                    v
+                                    [(b/read-buffer (l/k kv) k-type) v]))
+                    (recur))
+                  holder))))
+          (catch Throwable e
+            (raise "Fail to sample-kv: " e
+                   {:dbi dbi-name :n n :k-type k-type :v-type v-type}))
+          (finally
+            (if (l/read-only? rtx)
+              (l/return-cursor dbi cur)
+              (l/close-cursor dbi cur))
+            (when-not (l/writing? lmdb)
+              (i/return-rtx lmdb rtx))))))))
+
 (defmacro scan
   ([call error]
    `(scan ~call ~error false))